backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
websockets
google-generativeai
motor
pydantic-settings